+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased](https://github.com/logiek/http-status-codes/compare/1.1.2...master)
 
 ### Added
+- Added function to find a status code by a given reason phrase.
 - Added function to retrieve all status codes with their corresponding reason phrase.
 
 ## [1.1.2 (2021-10-28)](https://github.com/logiek/http-status-codes/compare/1.1.1...1.1.2)